🧬 Move all the base classes about dividing feature to new module *_divide*.
